fix(proxy): split forward.rs test module into a submodule (file-size gate)
The #160 (slow-backend warn) and #162 (X-Forwarded-For) changes grew
forward.rs to 599 lines, past CI's 500-line file-size gate — turning main
red after those merges landed. Move the `#[cfg(test)] mod tests` block into
`forward/tests.rs` (a submodule of the same module, `use super::*` unchanged),
dropping forward.rs to 376 lines. No production code moves; behavior identical.
